Circle CI updates
Replaces `assembleDebug` with `compileDebugSources` on Circle CI.

This way we can still detect compile errors in the sample app but avoid the dreaded `transformClassesWithDexForDebug` which spawns a ridiculous number of threads (4 per library) and eats up all the memory allocated to the container causing the build to intermittently fail. Nom Nom.

* Adds `compileDebugSources` to Gradle verify task
* Removes `assembleDebug` from `circle.yml` test tasks